+++ nxpo_hrms/nxpo_hrms/report/pnd1a_for_nxpo/pnd1a_for_nxpo.js
@@ -1,7 +1,7 @@
 // Copyright (c) 2016, Frappe Technologies Pvt. Ltd. and contributors
 // For license information, please see license.txt
 
-frappe.query_reports["PND1a NXPO"] = {
+frappe.query_reports["PND1a For NXPO"] = {
 	filters: [
 		{
 			fieldname: "company",
